User, UserDaoImpl, UserController: исправил замечания ментора.

+++ src/main/java/web/dao/UserDaoImpl.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Repository;
 import web.model.User;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
@@ -11,11 +12,11 @@ import java.util.List;
 public class UserDaoImpl implements UserDao {
 
     @PersistenceContext
-    EntityManager entityManager;
+    private EntityManager entityManager;
 
     @Override
     public void addUser(User user) {
-        entityManager.merge(user);
+        entityManager.persist(user);
     }
 
     @Override
@@ -24,23 +25,31 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User listUserId(int id) {
+    public User getUserById(int id) {
+        if (entityManager.find(User.class, id) == null) {
+            throw new EntityNotFoundException("Entity not found for given ID " + id);
+        }
         return entityManager.find(User.class, id);
     }
 
     @Override
     public void deleteUserById(int id) {
+        if (entityManager.find(User.class, id) == null) {
+            throw new EntityNotFoundException("Entity not found for given ID " + id);
+        }
         entityManager.remove(entityManager.find(User.class, id));
     }
 
     @Override
     public void editUser(int id, User userNew) {
+        if (entityManager.find(User.class, id) == null) {
+            throw new EntityNotFoundException("Entity not found for given ID " + id);
+        }
         User user = entityManager.find(User.class, id);
         if (user != null) {
             user.setName(userNew.getName());
             user.setLastname(userNew.getLastname());
             user.setAge(userNew.getAge());
-            entityManager.merge(user);
         }
     }
 }
